Added health to the project
Added health to the project

+++ Server1/Services/DistributionService/DistributionService.cs
@@ -20,27 +20,34 @@ public class DistributionService : IDistributionService
         _tcpService = tcpService;
     }
 
-    public async Task<IDictionary<int, Data>?> GetAll()
+    public async Task<KeyValuePair<int, Data>?> GetById(int id)
     {
-        var resultDictionary = await _dataService.GetAll();
+        //try get from local storage if not get from clusters
+        var res = await _dataService.GetById(id);
 
-        var server1Data = await _httpService.GetAll(Settings.Server1);
-
-        if (resultDictionary != null && server1Data != null)
+        if (Settings.Leader)
         {
-            foreach (var data in server1Data)
+            if (res.Value.Value == null)
             {
-                if (!resultDictionary.ContainsKey(data.Key))
-                {
-                    resultDictionary.Add(data);
-                }
+                res = await _httpService.GetById(id, Settings.Server2);
             }
         }
-        else if (resultDictionary != null)
+
+        return res;
+    }
+
+    public async Task<IDictionary<int, Data>?> GetAll()
+    {
+        //call get all from all servers and remove duplicates
+
+        var resultDictionary = await _dataService.GetAll();
+
+        if (Settings.Leader)
         {
+            //try to get from server 1 if not get from server 2
             var server2Data = await _httpService.GetAll(Settings.Server2);
 
-            if (server2Data != null)
+            if (resultDictionary != null && server2Data != null)
             {
                 foreach (var data in server2Data)
                 {
@@ -55,47 +62,28 @@ public class DistributionService : IDistributionService
         return resultDictionary;
     }
 
-    public async Task<KeyValuePair<int, Data>?> GetById(int id)
+    public async Task<Data> Update(int id, Data data)
     {
-        var data = await _dataService.GetById(id);
-        if (data?.Value == null)
+        //update try update all servers
+        if (Settings.Leader)
         {
-            data = await _httpService.GetById(id, Settings.Server1);
+            var server2Data = await _httpService.Update(id, data, Settings.Server2);
         }
 
-        if (data?.Value == null)
-        {
-            data = await _httpService.GetById(id, Settings.Server2);
-        }
-
-        return data;
+        return await _dataService.Update(id, data);
     }
 
     public async Task<IList<Result>> Save(Data data)
     {
         var results = new List<Result>();
 
-        var optimalServerNames = StorageHelper.GetOptimalServers();
+        var result = await _dataService.Save(data);
+        result.UpdateServerStatus();
 
-        if (optimalServerNames.Contains(Settings.ServerName))
-        {
-            var result = await _dataService.Save(data);
-            result.UpdateServerStatus();
+        results.Add(result);
 
-            results.Add(result);
-        }
-
-        if (optimalServerNames.Contains(ServerName.Server1))
-        {
-            var server1Response = _tcpService.TcpSave(data, Settings.Server1TcpSavePort);
-            if (server1Response != null)
-            {
-                server1Response.UpdateServerStatus();
-                results.Add(server1Response);
-            }
-        }
-
-        if (optimalServerNames.Contains(ServerName.Server2))
+        //use tcp to save data to other servers
+        if (Settings.Leader)
         {
             var server2Response = _tcpService.TcpSave(data, Settings.Server2TcpSavePort);
             if (server2Response != null)
@@ -108,38 +96,20 @@ public class DistributionService : IDistributionService
         return results;
     }
 
-    public async Task<Data> Update(int id, Data data)
-    {
-        await _httpService.Update(id, data, Settings.Server1);
-        await _httpService.Update(id, data, Settings.Server2);
-
-        return await _dataService.Update(id, data);
-    }
-
     public async Task<IList<Result>> Delete(int id)
     {
         var results = new List<Result>();
         var result = await _dataService.Delete(id);
         result.UpdateServerStatus();
 
-        var server1Result = await _httpService.Delete(id, Settings.Server1);
-        var server2Result = await _httpService.Delete(id, Settings.Server2);
-
-        if (server1Result != null)
+        if (Settings.Leader)
         {
-            foreach (var resultSummary in server1Result)
-            {
-                resultSummary.UpdateServerStatus();
-                results.Add(resultSummary);
-            }
-        }
+            var server2Result = await _httpService.Delete(id, Settings.Server2);
 
-        if (server2Result != null)
-        {
-            foreach (var resultSummary in server2Result)
+            if (server2Result != null)
             {
-                resultSummary.UpdateServerStatus();
-                results.Add(resultSummary);
+                server2Result.UpdateServerStatus();
+                results.Add(server2Result);
             }
         }
